requirements.txt
flask>=1.0
slack-bolt
gunicorn
python-dotenv
wheel
pyodbc
